Check user input randomness double range
AGenerator got a new method showWarning with a string input. The string
gets shown as a warning in showMessageDialog.

At the start of the generators run method user input randomness gets
checked for numberformatexceptions and for allowed 0.0 to 1.0 range. If
input throws a numberformatexceptions or is out of range showWarning
gets shown and run skips its normal generator calculating.

Wolfram, NaSch and GOL generators have randomness input and got a
modified run method.