Fix color models
Old 4 color models got reduced to three by combining Atari with RGB into
RGB. All work now as intended instead of just Atari.

Black-white model got new calculations that show the underlying picture
more by now using the RGB calculations as the starting point. They now
calculate the average of the red, green and blue values to set RGB to
the new average to get a black and white picture.

White noise multiplicates the red, green and blue values with the same
random number in the 0.0 to <1.0 range. This introduces a random element
in the picture which looks like white noise.

Simplified checks recursiveColorValCalc returns a value in 0.0 to 1.0
range.

Removed some vars that were not used in RandomTreeArt and
TestRandomTreeArt and their corresponding SideBar class.