Add load image to RandomTreeArt
Class MainWindow got a menu entry "Load Image as PNG" under file to load
a picture name from disk and parse from this name the values that it
used before. The menu entry "Save Image as PNG" now automatically adds
a name it gets from the method getFilePath in class IGenerator. It
returns an empty string by default. RTree classes override this and
return their generator settings width, height, generations and seed.
Which  then get stored in the file name.

Class IGenerator also got a setLoadedValues function which by default is
empty. RTree generators override it to split the filepath string into an
 array with gen name, width, height, generations, seed and file ending
 values. These then get modified and put into their respective GUI
 elements.

Class SideBarRTree has a setter for cbSeed to activate this field when
loading values with new menu entry. So the user does not have to click
it themselves.